add get incident and split out incident types into separate file
